fix: growthStep, SubscriberStep, PredictionViewStep 컴포넌트 위치 수정 및 레이아웃 수정
